Refactoring/SearchGame verschoben & navbar entfernt
